functions/data_events/requirements.txt
attrs==19.3.0
boto3==1.9.220
botocore==1.12.220
certifi==2019.6.16
chardet==3.0.4
coverage==5.2
docutils==0.15.2
idna==2.8
importlib-metadata==1.7.0
jmespath==0.9.4
kafka-python==1.4.7
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==0.24.0
pluggy==0.13.1
psycopg2==2.8.3
py==1.9.0
pycrypto==2.6.1
pydash==4.7.5
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-dateutil==2.8.0
python-dotenv==0.14.0
python-snappy==0.5.4
pytz==2020.1
PyYAML==5.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.6
urllib3==1.25.3
wcwidth==0.2.5
xmltodict==0.12.0
zipp==3.1.0
